refactor: refactor conversation module
