+++ src/conversation/conversation.service.ts
@@ -1,5 +1,5 @@
-import { MessageConversation } from './dto/message-conversation.dto';
-import { ProfileConversationDto } from './dto/profile-conversation.dto';
+import { MessageService } from './../message/message.service';
+import { ProfileConversationDto } from '../profile/dto/profile-conversation.dto';
 import { ProfileService } from './../profile/profile.service';
 import { Mapper } from '@automapper/core';
 import { InjectMapper } from '@automapper/nestjs';
@@ -11,27 +11,22 @@ import {
 } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { THttpResponse } from 'src/shared/common/http-response.dto';
-import { DataSource, Repository, Brackets } from 'typeorm';
+import { Repository } from 'typeorm';
 import { ConversationDto } from './dto/conversation.dto';
 import { CreateConversationDto } from './dto/create-conversation.dto';
 import { Conversation } from './entities/conversation.entity';
 import { MatchService } from '../match/match.service';
 import { CreateMatchDto } from '../match/dto/create-match.dto';
-import { Profile } from '../profile/entities/profile.entity';
 import { MessageGateway } from '../message/message.gateway';
-import {
-  IConversation,
-  IConversationSocket,
-  IConversationProfile,
-} from './interface';
-import { Message } from 'src/message/entities/message.entity';
+import { IConversation, IConversationSocket } from '../profile/interfaces';
+import { MessageConversation } from 'src/message/dto/message-conversation.dto';
 
 @Injectable()
 export class ConversationService {
   constructor(
     @InjectRepository(Conversation)
     private readonly _conversationRepository: Repository<Conversation>,
-    private _dataSource: DataSource,
+    private readonly _messageService: MessageService,
     private readonly _matchService: MatchService,
     private readonly _profileService: ProfileService,
     @InjectMapper() private readonly _classMapper: Mapper,
@@ -101,65 +96,6 @@ export class ConversationService {
     }
   }
 
-  async getConversationProfile(otherUserIds: IConversationProfile[]) {
-    try {
-      const profiles = await this._dataSource.manager
-        .createQueryBuilder(Profile, 'pro')
-        .where('pro.userId IN (:...userIds)', {
-          userIds: otherUserIds.map((u) => u.userId),
-        })
-        .getMany();
-
-      if (profiles.length === 0) {
-        throw new NotFoundException('Not Found Conversations');
-      }
-
-      const mapProfile = await this._classMapper.mapArrayAsync(
-        profiles,
-        Profile,
-        ProfileConversationDto,
-      );
-
-      return mapProfile;
-    } catch (err) {
-      throw new BadRequestException('Error Get Conversation Profiles');
-    }
-  }
-
-  async getLastMessage(otherUserIds: IConversationProfile[]) {
-    try {
-      const lastMessages = await Promise.all(
-        otherUserIds.map(async (u) => {
-          return await this._dataSource
-            .createQueryBuilder(Message, 'mess')
-            .where(
-              new Brackets((query) =>
-                query.where('mess.conversationId = :conversationId', {
-                  conversationId: u.conversationId,
-                }),
-              ),
-            )
-            .orderBy('mess.updatedAt', 'DESC')
-            .getOne();
-        }),
-      );
-
-      if (!lastMessages) {
-        throw new NotFoundException('Not Found Messages');
-      }
-
-      const mapLastMessages = await this._classMapper.mapArrayAsync(
-        lastMessages,
-        Message,
-        MessageConversation,
-      );
-
-      return mapLastMessages;
-    } catch (err) {
-      throw new BadRequestException('Error Get Last Message');
-    }
-  }
-
   async getConversationsById(userId: string): Promise<
     THttpResponse<
       {
@@ -190,8 +126,8 @@ export class ConversationService {
       });
 
       const [mapLastMessages, mapProfile] = await Promise.all([
-        await this.getLastMessage(otherUserIds),
-        await this.getConversationProfile(otherUserIds),
+        await this._messageService.getLastMessage(otherUserIds),
+        await this._profileService.getConversationProfile(otherUserIds),
       ]);
 
       const result = otherUserIds.map((u) => {
@@ -200,11 +136,7 @@ export class ConversationService {
           createdAt: u.createdAt,
           lastMessage: mapLastMessages.find((i) => {
             if (i.conversationId === u.conversationId) {
-              return {
-                content: i.content,
-                senderId: i.senderId,
-                type: i.type,
-              };
+              return i;
             }
           }),
           userProfile: mapProfile.find((p) => {
@@ -214,8 +146,6 @@ export class ConversationService {
           }),
         };
       });
-
-      console.log(result);
 
       return {
         statusCode: HttpStatus.OK,
