+++ src/message/dto/message-conversation.dto.ts
@@ -1,0 +1,19 @@
+import { AutoMap } from '@automapper/classes';
+import { MessageType } from 'src/shared/enums';
+
+export class MessageConversation {
+  @AutoMap()
+  conversationId: string;
+
+  @AutoMap()
+  content: string;
+
+  @AutoMap()
+  senderId: string;
+
+  @AutoMap()
+  type: MessageType;
+
+  @AutoMap()
+  createdAt: Date;
+}
