+++ src/message/message.gateway.ts
@@ -14,7 +14,7 @@ import * as dotenv from 'dotenv';
 import { GatewayGuard } from '../auth/guards/gateway.guard';
 import { MessageService } from './message.service';
 import { MessageType } from 'src/shared/enums';
-import { IConversationSocket } from '../conversation/interface';
+import { IConversationSocket } from '../shared/interfaces/conversation-profile.interface';
 
 dotenv.config();
