+++ src/message/message.module.ts
@@ -12,6 +12,6 @@ import { JwtModule } from '@nestjs/jwt';
   imports: [TypeOrmModule.forFeature([Message]), CloudinaryModule, JwtModule],
   controllers: [MessageController],
   providers: [MessageService, MessageGateway, MessageProfile],
-  exports: [MessageGateway],
+  exports: [MessageGateway, MessageService],
 })
 export class MessageModule {}
