+++ src/message/message.service.ts
@@ -2,10 +2,11 @@ import {
   BadRequestException,
   HttpStatus,
   Injectable,
+  NotFoundException,
   UploadedFile,
 } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { Repository } from 'typeorm';
+import { Brackets, DataSource, Repository } from 'typeorm';
 import { CreateMessageDto } from './dto/create-message.dto';
 import { Message } from './entities/message.entity';
 import { Server } from 'socket.io';
@@ -13,8 +14,10 @@ import { MessageDto } from './dto/message-dto';
 import { THttpResponse } from 'src/shared/common/http-response.dto';
 import { Mapper } from '@automapper/core';
 import { InjectMapper } from '@automapper/nestjs';
-import { CloudinaryService } from 'src/cloudinary/cloudinary.service';
-import { MessageType } from 'src/shared/enums';
+import { CloudinaryService } from '../cloudinary/cloudinary.service';
+import { MessageType } from '../shared/enums';
+import { IConversationProfile } from '../profile/interfaces';
+import { MessageConversation } from './dto/message-conversation.dto';
 
 @Injectable()
 export class MessageService {
@@ -23,6 +26,7 @@ export class MessageService {
     private readonly _messageRepository: Repository<Message>,
     @InjectMapper() private readonly _classMapper: Mapper,
     private readonly _cloudinaryService: CloudinaryService,
+    private readonly _dataSource: DataSource,
   ) {}
 
   async sendMessage(
@@ -98,6 +102,40 @@ export class MessageService {
       };
     } catch (err) {
       throw new BadRequestException(err.message);
+    }
+  }
+
+  async getLastMessage(otherUserIds: IConversationProfile[]) {
+    try {
+      const lastMessages = await Promise.all(
+        otherUserIds.map(async (u) => {
+          return await this._dataSource
+            .createQueryBuilder(Message, 'mess')
+            .where(
+              new Brackets((query) =>
+                query.where('mess.conversationId = :conversationId', {
+                  conversationId: u.conversationId,
+                }),
+              ),
+            )
+            .orderBy('mess.updatedAt', 'DESC')
+            .getOne();
+        }),
+      );
+
+      if (!lastMessages) {
+        throw new NotFoundException('Not Found Messages');
+      }
+
+      const mapLastMessages = await this._classMapper.mapArrayAsync(
+        lastMessages,
+        Message,
+        MessageConversation,
+      );
+
+      return mapLastMessages;
+    } catch (err) {
+      throw new BadRequestException('Error Get Last Message');
     }
   }
 }
