+++ src/profile/profile.service.ts
@@ -1,13 +1,20 @@
 import { THttpResponse } from './../shared/common/http-response.dto';
-import { BadRequestException, HttpStatus, Injectable } from '@nestjs/common';
+import {
+  BadRequestException,
+  HttpStatus,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { Repository } from 'typeorm';
+import { DataSource, Repository } from 'typeorm';
 import { CreateProfileDto } from './dto/create-profile.dto';
 import { UpdateProfileDto } from './dto/update-profile.dto';
 import { Profile } from './entities/profile.entity';
 import { InjectMapper } from '@automapper/nestjs';
 import { Mapper } from '@automapper/core';
 import { ProfileDto } from './dto/profile.dto';
+import { IConversationProfile } from './interfaces';
+import { ProfileConversationDto } from './dto/profile-conversation.dto';
 
 @Injectable()
 export class ProfileService {
@@ -15,6 +22,7 @@ export class ProfileService {
     @InjectRepository(Profile)
     private readonly _profileRepository: Repository<Profile>,
     @InjectMapper() private readonly _classMapper: Mapper,
+    private _dataSource: DataSource,
   ) {}
 
   async create(createProfileDto: CreateProfileDto) {
@@ -46,19 +54,6 @@ export class ProfileService {
 
   async findOneByUserId(userId: string): Promise<THttpResponse<ProfileDto>> {
     try {
-      // const profile = await this._profileRepository.findOne({
-      //   where: { userId },
-      // });
-
-      // console.log(profile);
-
-      // if (!profile) {
-      //   throw new HttpException(
-      //     'No profile found with that id!',
-      //     HttpStatus.NOT_FOUND,
-      //   );
-      // }
-
       const result = await this._classMapper.mapAsync(
         await this._profileRepository.findOne({
           where: { userId },
@@ -81,31 +76,38 @@ export class ProfileService {
     updateProfileDto: UpdateProfileDto,
   ): Promise<THttpResponse<void>> {
     try {
-      // const toUpdateDto = this._classMapper.map(
-      //   updateProfileDto,
-      //   UpdateProfileDto,
-      //   ProfileDto,
-      // );
-
       await this._profileRepository.update({ userId }, updateProfileDto);
-
-      // console.log(toUpdateDto);
-
-      // console.log(await this._profileRepository.save(toUpdateDto));
-
-      // const updatedProfile = this._classMapper.mapAsync(
-      //   await this._profileRepository.save(
-      //     Object.assign(profile.data, updateProfileDto),
-      //   ),
-      //   Profile,
-      //   ProfileDto,
-      // );
 
       return {
         statusCode: HttpStatus.NO_CONTENT,
       };
     } catch (err) {
       throw new BadRequestException(err.message);
+    }
+  }
+
+  async getConversationProfile(otherUserIds: IConversationProfile[]) {
+    try {
+      const profiles = await this._dataSource.manager
+        .createQueryBuilder(Profile, 'pro')
+        .where('pro.userId IN (:...userIds)', {
+          userIds: otherUserIds.map((u) => u.userId),
+        })
+        .getMany();
+
+      if (profiles.length === 0) {
+        throw new NotFoundException('Not Found Conversations');
+      }
+
+      const mapProfile = await this._classMapper.mapArrayAsync(
+        profiles,
+        Profile,
+        ProfileConversationDto,
+      );
+
+      return mapProfile;
+    } catch (err) {
+      throw new BadRequestException('Error Get Conversation Profiles');
     }
   }
 }
